+++ components/buttons/trash-button.tsx
