+++ components/navbar.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import UserHeader from "@/components/userHeader"
+import UserHeader from "@/components/user-header"
 import Image from "next/image";
 import Link from "next/link";
 import { useAuth } from "@/contexts/authContext";
