+++ components/petCard/card-image.tsx
