+++ components/petCard/pet-card.tsx
@@ -1,21 +1,15 @@
 'use client'
 import React, { useState } from "react";
 import clsx from "clsx";
-import CardImage from "@components/petCard/cardImage";
-import CardText from "./cardText";
-import FavoriteButton from "../buttons/FavoriteButton";
+import CardImage from "@/components/petCard/card-image";
+import CardText from "./card-text";
+import FavoriteButton from "../buttons/favorite-button";
+import { Post } from "@/types/post";
 
 //Defini estos tipos para que el componente no tenga errores, esto debera cambiar en el futuro cuando el endpoint que conecta
 //posts con pets este implementado
 
 
-type Post = {
-    id: string;
-    content: string;
-    date: string;
-    urlPhoto: string
-    imageUrl: string;
-};
 
 type PetCardProps = {
     post: Post;
@@ -46,7 +40,7 @@ export default function PetCard({ post, className }: PetCardProps) {
         <div className={clsx("w-64 rounded-xl overflow-hidden bg-white drop-shadow-md flex flex-col relative", className)}>
             <FavoriteButton variant={isFavorite ? "active" : "desactivated"} // Usa el estado para cambiar el 'variant'
                 onClick={() => setIsFavorite(!isFavorite)} className="absolute top-2 right-2 z-10" />
-            <CardImage image={post?.imageUrl} />
+            <CardImage image={post?.imageUrl || ""} />
             <CardText post={post} />
         </div>
     );
