+++ components/title.tsx
@@ -1,7 +1,7 @@
 import clsx from "clsx";
 import Link from "next/link";
 import React from "react";
-import { titleColor, titleText } from "../types/titles"
+import { titleColor, titleText } from "../types/title"
 
 interface TitleProps {
     title?: string;
