+++ types/post-type.ts
